+++ src/aks-preview/azext_aks_preview/_format.py
@@ -62,10 +62,18 @@ def aks_machine_show_table_format(result):
         entry["ipv4"] = ipv4_addresses
         entry["ipv6"] = ipv6_addresses
         parsed = compile_jmes("""{
-                name: name,
-                ipv4: ipv4,
-                ipv6: ipv6
-            }""")
+            name: name,
+            zones: zones,
+            ipv4: ipv4,
+            ipv6: ipv6,
+            nodeImageVersion: nodeImageVersion,
+            provisioningState: provisioningState,
+            orchestratorVersion: orchestratorVersion,
+            currentOrchestratorVersion: currentOrchestratorVersion,
+            vmSize: vmSize,
+            priority: priority,
+            mode: mode
+        }""")
         return parsed.search(entry, Options(dict_cls=OrderedDict))
     return parser(result)