+++ src/fleet/azext_fleet/constants.py
@@ -8,6 +8,9 @@ UPGRADE_TYPE_FULL = "Full"
 UPGRADE_TYPE_NODEIMAGEONLY = "NodeImageOnly"
 FLEET_1P_APP_ID = "609d2f62-527f-4451-bfd2-ac2c7850822c"
 
+SUPPORTED_GATE_STATES_FILTERS = ["Pending", "Skipped", "Completed"]
+SUPPORTED_GATE_STATES_PATCH = ["Completed"]
+
 UPGRADE_TYPE_ERROR_MESSAGES = {
     UPGRADE_TYPE_CONTROLPLANEONLY: f"Please set kubernetes version when upgrade type is '{UPGRADE_TYPE_CONTROLPLANEONLY}'.",  # pylint: disable=line-too-long
     UPGRADE_TYPE_FULL: f"Please set kubernetes version when upgrade type is '{UPGRADE_TYPE_FULL}'.",