+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/tap/_resync.py
@@ -17,7 +17,7 @@ from azure.cli.core.aaz import *
 class Resync(AAZCommand):
     """Implements the operation to the underlying resources.
 
-    :example: Resync Operation
+    :example: Resync the Network Tap
         az networkfabric tap resync --resource-group "example-rg" --network-tap-name "example-networktap"
     """
 
@@ -46,7 +46,7 @@ class Resync(AAZCommand):
 
         _args_schema = cls._args_schema
         _args_schema.network_tap_name = AAZStrArg(
-            options=["--network-tap-name"],
+            options=["--resource-name", "--network-tap-name"],
             help="Name of the Network Tap.",
             required=True,
             id_part="name",
@@ -222,12 +222,15 @@ class _ResyncHelper:
         additional_info.Element = AAZObjectType()
 
         _element = _schema_error_detail_read.additional_info.Element
-        _element.info = AAZFreeFormDictType(
+        _element.info = AAZDictType(
             flags={"read_only": True},
         )
         _element.type = AAZStrType(
             flags={"read_only": True},
         )
+
+        info = _schema_error_detail_read.additional_info.Element.info
+        info.Element = AAZAnyType()
 
         details = _schema_error_detail_read.details
         details.Element = AAZObjectType()