+++ src/palo-alto-networks/HISTORY.rst
@@ -2,6 +2,14 @@
 
 Release History
 ===============
+1.1.2b2
+++++++
+* Further Updates to CLI descriptions.
+
+1.1.2b1
+++++++
+* updated CLI descriptions.
+
 1.1.1b1
 ++++++
 * Fix #27401 `az palo-alto cloudngfw local-rulestack commit` command failing.